sv_move.c: add explicit int cast to a float argument to abs().
Previously, the cast was implicit since abs() takes an int,
so there is no change in behaviour.
Suppresses a clang warning that we might want fabs().

git-svn-id: https://svn.code.sf.net/p/quakespasm/code/trunk@1451 af15c1b1-3010-417e-b628-4374ebc0bcbd

+++ Quake/sv_move.c
@@ -320,7 +320,7 @@ void SV_NewChaseDir (edict_t *actor, edict_t *enemy, float dist)
 	}
 
 // try other directions
-	if ( ((rand()&3) & 1) ||  abs(deltay)>abs(deltax))
+	if ( ((rand()&3) & 1) ||  abs((int)deltay)>abs((int)deltax)) // ericw -- explicit int cast to suppress clang suggestion to use fabsf
 	{
 		tdir=d[1];
 		d[1]=d[2];
